pyproject.toml
[project]
name = "simpleharness"
version = "0.1.0"
description = "Lightweight baton-pass agent harness over the Claude Code CLI"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Ole Johan Bondahl" }]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
# NOTE: deal is a runtime dependency — core.py imports it for @deal.pure.
# Do NOT move it to [dependency-groups] dev. That breaks `uv tool install`.
dependencies = [
    "deal>=4.24.6",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.urls]
Homepage = "https://github.com/OleJBondahl/SimpleHarness"
Repository = "https://github.com/OleJBondahl/SimpleHarness"
Documentation = "https://github.com/OleJBondahl/SimpleHarness/blob/master/docs/usage.md"
Issues = "https://github.com/OleJBondahl/SimpleHarness/issues"

[project.scripts]
simpleharness = "simpleharness.shell:main"
simpleharness-approver-hook = "simpleharness.approver_shell:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["simpleharness"]

[tool.setuptools.package-data]
simpleharness = [
    "config.yaml",
    "simpleharness_approver_hook.sh",
    "roles/*.md",
    "subagents/*.md",
    "workflows/*.md",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
extend-select = [
    "I",      # isort — import ordering
    "B",      # flake8-bugbear — common gotchas
    "UP",     # pyupgrade — modernize syntax
    "SIM",    # flake8-simplify — simplifiable code
    "RUF",    # ruff-specific rules
    "D",      # pydocstyle — docstrings
    "S",      # flake8-bandit — security (replaces standalone bandit)
    "C90",    # mccabe — cyclomatic complexity
    "C4",     # flake8-comprehensions — unnecessary comprehensions
    "A",      # flake8-builtins — shadowing builtins
    "T20",    # flake8-print — no print() in production code
    "PT",     # flake8-pytest-style — pytest best practices
    "TID",    # flake8-tidy-imports — import hygiene
    "PERF",   # perflint — performance anti-patterns
    "FURB",   # refurb — modern Python idioms
    "LOG",    # flake8-logging — logging best practices
    "N",      # pep8-naming — naming conventions
    "TCH",    # flake8-type-checking — TYPE_CHECKING imports
]
extend-ignore = [
    "D100",   # missing docstring in public module
    "D104",   # missing docstring in public package
    "D105",   # missing docstring in magic method
    "D107",   # missing docstring in __init__
    "S101",   # assert used — we use assert in tests and deal contracts
    "S608",   # false positive — f-strings in prompt templates, not SQL
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
# Shell modules: allow print, relax complexity
"src/simpleharness/shell.py" = ["T20", "C90", "S603", "S607"]
"src/simpleharness/approver_shell.py" = ["T20", "C90", "S603", "S607"]
"src/simpleharness/session.py" = ["T20", "C90", "S603", "S607"]
"src/simpleharness/io.py" = ["T20", "C90"]
"src/simpleharness/process.py" = ["T20", "S603", "S607"]
"src/simpleharness/benchmark.py" = ["T20", "C90", "S603", "S607"]
"src/simpleharness/ui.py" = ["T20"]
# Tests: relax docstrings, asserts, print, security, naming
"tests/**" = ["D", "S", "T20", "PT", "N"]
# Scripts: relax docstrings, print, security
"scripts/**" = ["D", "T20", "S"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.9",
    "ty>=0.0.29",
    "hypothesis>=6.151.0",
    "polyfactory>=2.22.0",
    "mutmut>=3.4.0",
    "import-linter>=2.5",
    "vulture>=2.14",
    "complexipy>=3.3.0",
]
security = [
    "detect-secrets>=1.5.0",
    "pip-audit>=2.7.0",
    "pip-licenses>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=src/simpleharness --cov-branch --cov-fail-under=42"

[tool.coverage.run]
branch = true
source = ["src/simpleharness"]

[tool.coverage.report]
fail_under = 42
show_missing = true
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.importlinter]
root_packages = ["simpleharness"]

[[tool.importlinter.contracts]]
name = "Core must not import shell"
type = "forbidden"
source_modules = ["simpleharness.core", "simpleharness.approver_core"]
forbidden_modules = ["simpleharness.shell", "simpleharness.approver_shell"]

[[tool.importlinter.contracts]]
name = "Shell layers are independent"
type = "independence"
modules = [
    "simpleharness.shell",
    "simpleharness.approver_shell",
]

[tool.mutmut]
paths_to_mutate = ["src/"]
tests_dir = "tests/"

[tool.vulture]
min_confidence = 80
paths = ["src"]

[tool.complexipy]
paths = ["src"]
max_complexity_allowed = 15

[tool.ty.rules]
all = "error"
